Move `Maybe*` traits back to `iced_futures`
