+++ core/src/lib.rs
@@ -35,7 +35,6 @@ mod color;
 mod content_fit;
 mod element;
 mod length;
-mod maybe;
 mod padding;
 mod pixels;
 mod point;
@@ -60,7 +59,6 @@ pub use font::Font;
 pub use gradient::Gradient;
 pub use layout::Layout;
 pub use length::Length;
-pub use maybe::{MaybeSend, MaybeSync};
 pub use overlay::Overlay;
 pub use padding::Padding;
 pub use pixels::Pixels;
