+++ futures/src/event.rs
@@ -1,8 +1,8 @@
 //! Listen to runtime events.
 use crate::core::event::{self, Event};
 use crate::core::window;
-use crate::core::MaybeSend;
 use crate::subscription::{self, Subscription};
+use crate::MaybeSend;
 
 /// Returns a [`Subscription`] to all the ignored runtime events.
 ///
