+++ futures/src/executor.rs
@@ -1,5 +1,6 @@
 //! Choose your preferred executor to power a runtime.
-use crate::core::MaybeSend;
+use crate::MaybeSend;
+
 use futures::Future;
 
 /// A type that can run futures.
