+++ futures/src/lib.rs
@@ -8,6 +8,7 @@
 pub use futures;
 pub use iced_core as core;
 
+mod maybe;
 mod runtime;
 
 pub mod backend;
@@ -17,6 +18,7 @@ pub mod keyboard;
 pub mod subscription;
 
 pub use executor::Executor;
+pub use maybe::{MaybeSend, MaybeSync};
 pub use platform::*;
 pub use runtime::Runtime;
 pub use subscription::Subscription;
