+++ futures/src/maybe.rs
