+++ futures/src/subscription/tracker.rs
@@ -1,6 +1,5 @@
-use crate::core::MaybeSend;
 use crate::subscription::{Event, Hasher, Recipe};
-use crate::BoxFuture;
+use crate::{BoxFuture, MaybeSend};
 
 use futures::channel::mpsc;
 use futures::sink::{Sink, SinkExt};
